requirements.txt
# Hala AI Service Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Vector Database
chromadb>=0.4.22

# Machine Learning / NLP
sentence-transformers>=2.2.2
numpy>=1.24.0

# LLM Providers
google-generativeai>=0.3.2
openai>=1.10.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
